Fix first fold closes everything
By default the fold level is 0, but it is not immediately applied. So
what happens when you first fold is that it folds straight to 0, closing
everything.

By setting the foldlevel high, the first time you fold it will find the
current maximum fold level and just reduce it by one, which is what one
would expect from a fold function.